Media creation (#20)
* Configured Sentry.io

* Updated file structure

* Updated file structure

* Added an integration test for /users

* Updated controllers

* Updated models export file

* Added tests for users endpoint

* Updated travis ci env vars

* Update

* Updated docs for /users/register and updated tests

* Updated /users/register docs and added tests

* Created tests for createUser model

* Added tests for user models

* Updated models tests

* Updated /users/register, updated docs, and added tests

* Added createUser and retrieveUserBy models and added tests

* Added password hashing

* Updated retrieveUserBy model

* Added JWT to /users/register

* Updated app.json with jwt key for review apps

* Updated .travis.yml with jwt key for testing

* Test

* Test

* Test

* Added error handler for user routes

* Added error handler for user routes

* Updated .travis.yml

* Added /users/:user_id endpoint and added tests

* Updated user endpoints and API docs

* Added endpoint for logging in

* Added endpoint for logging in

* Updated API Docs

* Updated API Docs

* Udpated API docs and updated endpoint integration tests

* Refactored tests

* Refactored controller integration tests

* Created updateUser model and added tests

* Created edit user endpoint and added tests

* basic photo uploading

* Update knexfile.js

* Update knexfile.js

* Created remove user endpoint and added tests

* Updated API docs

* further cleaned up merge

* Added a test for deleteUserById model

* Added route protection/permissions with JWT and added endpoint tests

* Testing CORS options

* Testing CORS options

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Testing heroku timeout

* Fixed a bug

* Fixed a bug

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Testing cors

* Configured CORS for all environments

* Configured CORS for all environments

* Configured CORS for all environments

* Configured CORS for all environments

* Updated API docs

* Updated API docs

* Updated API docs

* Refactored controllers

* Created createAlbum model and added tests

* Created add meta data endpoint

* Added tests for add meta endpoint

* Added endpoint for editing an album and added tests

* Created remove album endpoint and added tests

* Created models and validation functions

* Resolving merge conflicts

* Updated endpoint tests

* Updated API docs and added tests

* Endpoint adds media to albums

* Endpoint adds media to albums

* Created endpoint for uploading media and added tests

* Updated create media endpoint

* Created endpoint for uploading media and added tests

* Added permission checking and tests for create media endpoint

* Added permission checking and tests for create media endpoint

* Added retrieveAlbumsMedia model and tests

* Created endpoint for getting an albums media and added tests

* Updated get albums media route description

* Updated get albums media route description

* Updated get albums media route description

* Added cover_url and thumbnail_url

* Updated album endpoints with cover_url

* Added get users media endpoint and tests

* refactored media upload

* added parsing IDs to integer

* testing cloudinary

* Added absolute path for tmp media folder

* Added .getkeep

* Updated utils

* Added bigSerial to user_id for postgresql

* Refactored migrations

+++ data/migrations/20191104154517_mediaAlbums.js
@@ -5,7 +5,7 @@ exports.up = function(knex) {
   return knex.schema.createTable('mediaAlbums', table => {
 
     // media_id
-    table.bigInteger('media_id')
+    table.integer('media_id')
       .notNullable()
       .references('media_id')
       .inTable('media')
@@ -13,7 +13,7 @@ exports.up = function(knex) {
       .onDelete('CASCADE');
 
     // album_id
-    table.bigInteger('album_id')
+    table.integer('album_id')
       .notNullable()
       .references('album_id')
       .inTable('albums')
